Add nock to tests and cloudinary support.

+++ src/controllers/meals.js
@@ -6,38 +6,90 @@ const Meal = require('../models/sequelize').meal;
 const User = require('../models/sequelize').user;
 const Like = require('../models/sequelize').like;
 const MealPic = require('../models/sequelize').meal_pic;
-// File system
-const fs = require('fs');
+const ProfilePic = require('../models/sequelize').profile_pic;
+// Cloudinary
+const cloudinary = require('cloudinary').v2;
+
+function deleteCloudinaryImage(public_id) {
+  return new Promise(function (resolve, reject) {
+    cloudinary.uploader.destroy(public_id, {
+      invalidate: true
+    }, function(error,result) {
+      if (error) reject(error);
+      resolve(result);
+    });
+  });
+};
 
 module.exports = {
 
-  getMeal: async (req, res) => {
+  getMealById: async (req, res) => {
     try {
       let meal = await Meal.findOne({
-        where: sequelize.where(
-          sequelize.fn('lower', sequelize.col('name')), 
-          sequelize.fn('lower', req.query.name)
-        ),
+        where: {
+          id: req.query.id
+        },
         include: [
-          { model: User, as: "creator", attributes: ['username']},
+          { 
+            model: User, 
+            as: "creator", 
+            attributes: ['username'],
+            include: [
+              { model: ProfilePic, as: 'profilePic', attributes: ['profilePicName'], duplicating: false }
+            ]
+          },
           { model: MealPic, as: 'mealPic', attributes: ['mealPicName'], duplicating: false },
+          { model: Like, attributes: ['userId'], duplicating: false }
         ]
       });
 
-      if (!meal) return res.status(404).json({ message: 'That meal does not exist.' });
+      if (!meal) return res.status(404).json({ message: 'That recipe does not exist.' });
 
       res.status(200).json(meal);
     } catch (err) {
-      res.status(500).json({ message: 'There was an error getting the meal.' });
+      res.status(500).json({ message: 'There was an error getting the recipe.' });
+    }
+  },
+
+  getMealByName: async (req, res) => {
+    try {
+      let mealName = req.query.name.replace('%20', ' ');
+
+      let meal = await Meal.findOne({
+        where: sequelize.where(
+          sequelize.fn('lower', sequelize.col('name')), 
+          sequelize.fn('lower', mealName)
+        ),
+        include: [
+          { 
+            model: User, 
+            as: "creator", 
+            attributes: ['username'],
+            include: [
+              { model: ProfilePic, as: 'profilePic', attributes: ['profilePicName'], duplicating: false }
+            ]
+          },
+          { model: MealPic, as: 'mealPic', attributes: ['mealPicName'], duplicating: false },
+          { model: Like, attributes: ['userId'], duplicating: false }
+        ]
+      });
+
+      if (!meal) return res.status(404).json({ message: 'That recipe does not exist.' });
+
+      res.status(200).json(meal);
+    } catch (err) {
+      res.status(500).json({ message: 'There was an error getting the recipe.' });
     }
   },
 
   findAvailableMealName: async (req, res) => {
     try {
+      let mealName = req.query.name.replace('%20', ' ');
+
       let meal = await Meal.findOne({
         where: sequelize.where(
           sequelize.fn('lower', sequelize.col('name')), 
-          sequelize.fn('lower', req.query.name)
+          sequelize.fn('lower', mealName)
         )
       });
 
@@ -47,7 +99,7 @@ module.exports = {
         res.status(400).json();
       }
     } catch (err) {
-      res.status(500).json({ message: 'There was an error finding available meal names.' });
+      res.status(500).json({ message: 'There was an error finding available recipe names.' });
     }
   },
 
@@ -55,34 +107,31 @@ module.exports = {
     try {
       req.body.creatorId = req.decoded.id;
       
+      // save these properties to variables
+      let tempMealPicName = req.body.mealPicName;
+      let tempPublicId = req.body.publicId;
+
+      // delete properties from req.body because the meal model
+      // does not have them
+      delete req.body.mealPicName;
+      delete req.body.publicId;
+
       let meal = await Meal.create(req.body);
 
-      if (req.file) {
+      if (tempMealPicName && tempPublicId) {
         let mealPic = await MealPic.create({
           mealId: meal.dataValues.id,
-          mealPicName: req.file.filename
+          mealPicName: tempMealPicName,
+          publicId: tempPublicId
         });
       }
 
       res.status(201).json({
         id: meal.id,
+        name: meal.name,
         message: "Meal successfully created."
       });
     } catch (err) {
-      // This deletes any meal picture a user uploaded during an attempt that had
-      // an error. This is deleted so if they change the meal name on any following 
-      // attempts, there won't be an unused picture.
-      if (req.file) {
-        // Checks if the meal picture exists
-        fs.stat(`public/images/mealPics/${req.file.filename}`, (err, stats) => {
-          if (stats) {
-            // Deletes profile picture
-            fs.unlink(`public/images/mealPics/${req.file.filename}`, (err) => {
-              if (err) return res.status(500).json({ message: 'There was an error.' });
-            });
-          }
-        });
-      }
 
       if (err.errors) {
         if (err.errors[0].message === 'name must be unique') {
@@ -96,56 +145,81 @@ module.exports = {
 
   update: async (req, res) => {
     try {
-      if (req.file) {
-        let previousMealPic = await MealPic.findOne({
-          where: { mealId: req.body.id }
-        });
-
-        // This deletes the old meal pic and creates the new one
-        if (previousMealPic) {
-          // Checks if the meal picture exists
-          fs.stat(`public/images/mealPics/${previousMealPic.dataValues.mealPicName}`, (err, stats) => {
-            if (stats) {
-              // Deletes meal picture
-              fs.unlink(`public/images/mealPics/${previousMealPic.dataValues.mealPicName}`, (err) => {
-                if (err) return res.status(500).json({ message: 'There was an error.' });
-              });
-            }
-          });
-        }
-
-        let mealPicToDelete = MealPic.destroy({
-          where: { mealId: req.body.id }
-        })
-
-        let mealPic = await MealPic.create({
-          mealId: req.body.id,
-          mealPicName: req.file.filename
-        });
-      }
-
-      let meal = await Meal.update(req.body, {
-        where: { 
+      let mealOwner = await Meal.findOne({
+        where: {
           id: req.body.id,
           creatorId: req.decoded.id
         }
       });
 
-      if (meal[0] === 1) {
+      if (!mealOwner) throw Error('You do not have permission to edit this recipe.');
+
+      let tempMealId = req.body.id;
+
+      delete req.body.id;
+
+      // save these properties to variables
+      let tempMealPicName = req.body.mealPicName;
+      let tempPublicId = req.body.publicId;
+
+      // delete properties from req.body because the meal model
+      // does not have them
+      delete req.body.mealPicName;
+      delete req.body.publicId;
+
+      let meal = await Meal.update(req.body, {
+        where: { 
+          id: tempMealId
+        }
+      });
+
+      let mealPic;
+      if (tempMealPicName && tempPublicId) {
+        let previousMealPic = await MealPic.findOne({
+          where: { mealId: tempMealId }
+        });
+
+        // This deletes the old meal pic from cloudinary
+        // result returns { result: 'ok' } on success
+        if (previousMealPic) {
+          let cloudImg = await deleteCloudinaryImage(tempPublicId);
+        }
+        
+        // If .upsert creates a new object, it returns true. If it updates
+        // an old object, it returns false. Passing {returning: true} returns the 
+        // updated object and make profilePic a truthy variable.
+        mealPic = await MealPic.upsert({
+          mealId: tempMealId,
+          mealPicName: tempMealPicName,
+          publicId: tempPublicId
+        }, {returning: true});
+      }
+
+      // If the user only updates the mealPic, then the user object will return 0 because nothing
+      // has updated. This if statement will allow a truthy result if the user only updated the mealPic
+      // or the meal object.
+      if (meal[0] === 1 || mealPic) {
         return res.status(201).json({ 
-          id: req.body.id,
-          message: 'Meal successfully updated.' 
+          id: tempMealId,
+          message: 'Recipe successfully updated.' 
         });
       } else {
         throw Error();
       }
     } catch (err) {
+      if (err.message === 'You do not have permission to edit this recipe.') {
+        return res.status(403).json({ message: err.message });
+      }
+      if (err.message === 'There was an error updating your recipe image.') {
+        return res.status(500).json({ message: err.message });
+      }
+
       if (err.errors) {
         if (err.errors[0].message === 'name must be unique') {
-          return res.status(400).json({ message: 'This meal name is already taken.' });
+          return res.status(400).json({ message: 'This recipe name is already taken.' });
         }
       }
-      res.status(500).json({ message: 'There was an error updating your meal.' });
+      res.status(500).json({ message: 'There was an error updating your recipe.' });
     }
   },
 
@@ -158,30 +232,21 @@ module.exports = {
         }
       });
 
+      if (!meal) return res.status(403).json({ message: 'You do not have permission to delete this recipe.' });
+
       let mealPicToDelete = await MealPic.findOne({
         where: { mealId: req.body.id }
       });
       
       if (mealPicToDelete) {
-        // Checks if the meal picture exists
-        fs.stat(`public/images/mealPics/${mealPicToDelete.dataValues.mealPicName}`, (err, stats) => {
-          if (stats) {
-            // Deletes meal picture
-            fs.unlink(`public/images/mealPics/${mealPicToDelete.dataValues.mealPicName}`, (err) => {
-              if (err) return res.status(500).json({ message: 'There was an error.' });
-            });
-          }
-        });
-
-        let deletePic = await MealPic.destroy({
-          where: { mealId: req.body.id }
-        })
+        let cloudImg = await deleteCloudinaryImage(mealPicToDelete.publicId);
       }
 
       if (meal) {
         let deletedRelatedData = await Promise.all([
           SavedMeal.destroy({where: { mealId: meal.dataValues.id }}),
-          Like.destroy({where: { mealId: meal.dataValues.id }})
+          Like.destroy({where: { mealId: meal.dataValues.id }}),
+          MealPic.destroy({where: { mealId: req.body.id }})
         ]);
       }
 
@@ -193,12 +258,12 @@ module.exports = {
       });
 
       if (deleted) {
-        return res.status(200).json({ message: 'Meal successfully deleted.' });
+        return res.status(200).json({ message: 'Recipe successfully deleted.' });
       } else {
         throw Error();
       }
     } catch (err) {
-      res.status(500).json({ message: 'There was an error deleting your meal.' });
+      res.status(500).json({ message: 'There was an error deleting your recipe.' });
     }
   },
 }
